code review suggestion : dead code cleanup
